Add dumb-fallback for loading parsers from scripts
Extends support on Python3 by side-stepping the issue with
None compatibility in the AST and try/except Python3 headers.

Scripts are loaded, cropped
before the `.parse_args(` call and imported as normal. The
result should be guaranteed to get the argparse object
without executing the subsequent script.

The calls to imp replace the script_name parameter with
`__main__` to accomodate the common convention of wrapping
execution in `if __name__ == '__main__':` (the script_name
is always __main__ while executing).