Implement heuristic ranking for parsers
Replacing the previous fixed requirement for extension and
file contents is a ranking approach which scores scripts against
a list of common contents, patterns and extensions. The result is
a number 0..1 reflecting how closely the script matches the
pattern. For most scripts this will result in the closest match
of parser being applied first to speed up parsing.

Possible improvements would be a base ranking value
for each parser to prefer the most common one when
two scripts score the same (e.g. always prefer argparse first).